major/pyproject.toml
[project]
name = "major"
version = "0.1.0"
description = "Major chat agent - Claude Agent SDK wrapper for interactive conversations"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "claude-agent-sdk>=0.1.18",
    "fastapi>=0.111.0",
    "uvicorn>=0.30.0",
    "pyyaml>=6.0",
    "python-frontmatter>=1.0.0",
    "pypdf>=4.0.0",
    "python-multipart>=0.0.9",
    "openai>=1.0.0",
    "anthropic>=0.40.0",
    "python-docx>=0.8.11",
]

[project.scripts]
major-server = "major.server:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/major"]
